2015/05/21 18:08

+++ worker/MainWorker.js
@@ -6,18 +6,26 @@ config.env().file({ "file":"config.json" });
 
 process.on('SIGTERM',function(){
 	console.log("\nGracefully shutting down from SIGTERM (system kill)");
-	singleton.getInstance().zkWorker.kill('SIGTERM');
-	process.exit();
+	if(singleton.getInstance().zkWorker){
+		singleton.getInstance().zkWorker.kill('SIGTERM');
+	}
+	process.exit(2);
 });
 
 process.on('SIGINT', function() {
 	//this.zkWorker.
 	//this.zkWorker.kill('SIGINT');
 	console.log("\nGracefully shutting down from SIGINT (Ctrl+C)");
-	process.exit();
+	if(singleton.getInstance().zkWorker){
+		singleton.getInstance().zkWorker.kill('SIGINT');
+	}
+	process.exit(1);
 });
 
 process.on('exit',function(code){
+	if(singleton.getInstance().zkWorker){
+		singleton.getInstance().zkWorker.exit(0);
+	}
 	console.log('Main process exited with exit code '+code);
 })
 
@@ -27,6 +35,8 @@ var singleton = function singleton(){
 	var env = process.env;
 	var	zkEnv = {};
 	var	child_process = require('child_process');
+	var zkWorker = null;
+	var workerHandler = {};
 
 	//Duplicate the parent's environment object
 	for (someVar in env) {
@@ -40,10 +50,18 @@ var singleton = function singleton(){
 	/*
 	 * send to zkWorker, message format: {'workerid':id,'action':action,'data':{}}
 	 */
-	this.zkWorker = child_process.fork('./worker/zkClientWorker.js',
+	zkWorker = child_process.fork('./worker/zkClientWorker.js',
 			{env:zkEnv}
 	);
 	
+	console.log("Zk Client Worker Start");
+	zkWorker.on('message', function(msg) {
+		  console.log('www Got message from zkWorker:', msg);
+	});
+	workerHandler['zkClient'] = zkWorker;
+	this.workerHandler = workerHandler;
+	console.log("Zk Client Worker initial.");
+
 	/*boss = child_process.spawn("node", ["worker/taskBoss.js"],
 			{env:bossEnv,stdio: ['ipc']}
 	);
@@ -51,24 +69,21 @@ var singleton = function singleton(){
 	boss.stdout.on('data', function(data) {
 	  console.log(data);
 	});*/
-
-	console.log("Zk Client Worker Start");
-	this.zkWorker.on('message', function(msg) {
-		  console.log('www Got message from zkWorker:', msg);
-	});
-	console.log("Zk Client Worker initial.");
 }
 
 singleton.prototype.sendMessage = function(msg){
-	this.zkWorker.send(msg);
+	console.log(msg['worker']);
+	this.workerHandler[msg['worker']].send(msg);
+//	this.zkWorker.send(msg);
 }
 
-singleton.prototype.getZKWorker = function(){
+/*singleton.prototype.getZKWorker = function(){
 	return this.zkWorker;
-}
+}*/
 
 singleton.instance = null;
-singleton.zkWorker = null;
+//singleton.zkWorker = null;
+singleton.workerHandler = {};
 /**
  * Singleton getInstance definition
  * @return singleton class
